vng/requirements.txt
asgiref==3.5.2
astor==0.8.1
async-generator==1.10
attrdict==2.0.1
attrs==22.1.0
Babel==2.10.3
backports.zoneinfo==0.2.1
bce-python-sdk==0.8.74
beautifulsoup4==4.11.1
boto3==1.24.93
botocore==1.27.93
bottle==0.12.23
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
common==0.1.2
contourpy==1.0.5
cssselect==1.1.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.32
data==0.4
decorator==5.1.1
deep-translator==1.9.0
dill==0.3.5.1
Django==4.1.2
django-cors-headers==3.13.0
django-storages==1.13.1
djangorestframework==3.14.0
dual==0.0.10
dynamo3==0.4.10
et-xmlfile==1.1.0
filelock==3.8.0
fire==0.4.0
Flask==2.2.2
Flask-Babel==2.0.0
flywheel==0.5.4
fonttools==4.37.4
funcsigs==1.0.2
future==0.18.2
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
h11==0.13.0
huggingface-hub==0.10.1
idna==3.4
imageio==2.22.2
imgaug==0.4.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
lanms-neo==1.0.2
lmdb==1.3.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.1
multiprocess==0.70.13
networkx==2.8.7
numpy==1.23.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
paddle==1.0.2
paddle-bfloat==0.1.7
paddleocr==2.6.0.1
paddlepaddle==2.4.0rc0
pandas==1.5.1
pdf2docx==0.5.6
peewee==3.15.3
Pillow==9.2.0
Polygon3==3.0.9.1
premailer==3.10.0
protobuf==3.20.0
prox==0.0.17
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyclipper==1.3.0.post3
pycparser==2.21
pycryptodome==3.15.0
PyMuPDF==1.19.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
python-geoip-python3==1.3
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
rapidfuzz==2.11.1
regex==2022.9.13
requests==2.28.1
s3transfer==0.6.0
scikit-image==0.19.3
scipy==1.9.2
selenium==4.4.3
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
termcolor==2.1.0
tifffile==2022.10.10
tight==0.1.0
tokenizers==0.13.1
torch==1.12.1
tqdm==4.64.1
transformers==4.23.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.5
urllib3==1.26.12
visualdl==2.4.1
webdriver-manager==3.8.3
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.10.0
